_Repからintへの変換にwarningが出ていたので、明示的なキャストを追加。

+++ libsources/internal/TimeDebugger/TimeDebugger.cpp
@@ -14,7 +14,7 @@ namespace debugTool
         void TimeDebugger::Stop()
         {
             auto end = GET_NOW;
-            elapsedMs = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
+            elapsedMs = (int)std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
         }
 
         void TimeDebugger::ToStream(std::ostream &os) const
